pyproject.toml
[project]
name = "alpha-machine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "supabase>=2.0.0",
    "slack-sdk>=3.0.0",
    "slack-bolt>=1.0.0",
    "flask>=2.0.0",
    "notion-client>=2.0.0"
]

[project.scripts]
alpha-machine = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
